TDD: Destroy owner of job_post

+++ app/controllers/job_posts_controller.rb
@@ -27,10 +27,13 @@ class JobPostsController < ApplicationController
         @job_posts=JobPost.all.order(created_at: :desc)
     end
     def destroy
+        
         job_post= JobPost.find params[:id]
-        job_post.destroy
-        flash[:danger]='Deleted job post'
-        redirect_to job_posts_path
+        if can?(:delete, job_post)
+            job_post.destroy
+            flash[:danger]='Deleted job post'
+            redirect_to job_posts_path
+        end
     end
     def edit
